Removed deprecated data_explorer.py module and modularized argparser
